Create UGE job array script to run Rscript commands from a supplied file
rscript-array.sh - Script to run a file of Rscript commands as an array job
It loads the R module and then runs a single line based on the TASK_ID
from a supplied file. Filename is the argument to the script.
The second argument is an optional job name for the success/failure message